requirements.txt
python-telegram-bot
ts3
pymysql
